rename internal field longValue und doubleValue

+++ src/main/kotlin/de/gma/gamma/builtins/FunctionTwoNumbersToBoolean.kt
@@ -18,13 +18,13 @@ abstract class FunctionTwoNumbersToBoolean() : BuiltinFunction(listOf("a", "b"))
         val b = extractNumber(callParams[1].evaluate(scope))
 
         if (a is IntegerValue && b is IntegerValue)
-            return BooleanValue.build(operationInteger(a.intValue, b.intValue))
+            return BooleanValue.build(operationInteger(a.longValue, b.longValue))
 
         if ((a is FloatValue || a is IntegerValue) && (b is FloatValue || b is IntegerValue)) {
             val af = a.toFloat()
             val bf = b.toFloat()
 
-            return BooleanValue.build(operationFloat(af.floatValue, bf.floatValue))
+            return BooleanValue.build(operationFloat(af.doubleValue, bf.doubleValue))
         }
 
         return null
